Add main menu. (#3)
This provides a main menu which allows the player to start a solo game, join a multiplayer game, or host one.  The `--autostart` flag can be used to start a solo game while bypassing the menu.  The `--client` and `--server` flags will also bypass the menu.

I wish Godot had something like React.